Additional comments and refactor
